# tmpl/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "MMock",
    "description": "Mmock is a testing and fast prototyping tool for developers",
    "license": {
      "name": "MIT",
      "url": "https://github.com/jmartin82/mmock/blob/master/LICENSE.md"
    }
  },
  "host": "localhost:8082",
...
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/request/verify": {
      "post": {
        "summary": "Verify requests",
        "tags": [
          "request"
        ],
        "operationId": "Verify requests",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Request"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of request",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/MockDefinition"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "MockDefinition": {
      "title": "Mock definition",
      "type": "object",
      "properties": {
        "URI": {
          "description": "",
          "type": "string"
        },
        "description": {
          "description": "",
          "type": "string"
        },
        "request": {
          "$ref": "#/definitions/Request"
        },
        "response": {
          "$ref": "#/definitions/Response"
        },
        "control": {
          "$ref": "#/definitions/Control"
        }
      },
      "required": [
        "URI",
        "description",
        "request",
        "response",
        "control"
      ]
    },
    "Request": {
      "title": "Request",
      "type": "object",
      "properties": {
        "host": {
          "description": "",
          "type": "string"
        },
        "method": {
          "description": "",
          "type": "string"
        },
        "path": {
          "description": "",
          "type": "string"
        },
        "queryStringParameters": {
          "$ref": "#/definitions/QueryStringParameters"
        },
        "headers": {
          "$ref": "#/definitions/KeyArray"
        },
        "cookies": {
          "type": "object",
          "description": "",
          "properties": {
            "key": {
              "type": "string"
            },
            "value": {
              "type": "string"
            }
          }
        },
        "body": {
          "description": "",
          "type": "string"
        }
      },
      "required": [
        "method",
        "path"
      ]
    },
    "QueryStringParameters": {
      "$ref": "#/definitions/KeyArray"
    },
    "Response": {
      "title": "Response",
      "type": "object",
      "properties": {
        "statusCode": {
          "description": "",
          "type": "integer",
          "format": "int32"
        },
        "headers": {
          "$ref": "#/definitions/KeyArray"
        },
        "cookies": {
          "type": "object",
          "description": "",
          "properties": {
            "key": {
              "type": "string"
            },
            "value": {
              "type": "string"
            }
          }
        },
        "body": {
          "description": "",
          "type": "string"
        }
      },
      "required": [
        "statusCode"
      ]
    },
    "KeyArray": {
      "title": "Values",
      "type": "object",
      "properties": {
        "key": {
          "description": "",
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "Control": {
      "title": "Control",
      "type": "object",
      "properties": {
        "priority": {
          "description": "",
          "type": "integer",
          "format": "int32"
        },
        "delay": {
          "description": "",
          "type": "integer",
          "format": "int32"
        },
        "crazy": {
          "description": "",
          "type": "boolean"
        },
        "scenario": {
          "$ref": "#/definitions/Scenario"
        },
        "proxyBaseURL": {
          "description": "",
          "type": "string"
        }
      }
    },
    "Scenario": {
      "title": "Scenario",
      "type": "object",
      "properties": {
        "name": {
          "description": "",
          "type": "string"
        },
        "requiredState": {
          "description": "",
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "newState": {
          "description": "",
          "type": "string"
        }
      }
    },
...
  }
}